pyproject.toml
[build-system]
requires = [
    "setuptools",
    # needed for building tables for the sdist:
    "PyYAML",
    "click",
]
build-backend = "setuptools.build_meta"

[project]
name = "Mathics_Scanner"
description = "Character Tables and Tokenizer for Mathics and the Wolfram Language."
dependencies = [
    "PyYAML",     # Used in mathics3-generate-json-table
    "chardet",    # Used in mathics_scanner.feed
    "click",      # Using in CLI: mathics3-generate-json-table
]
requires-python = ">=3.10"
readme = "README.rst"
license = "GPL-3.0-or-later"
keywords = ["Mathematica", "Wolfram", "Interpreter", "Shell", "Math", "CAS"]
maintainers = [
    {name = "Mathics3 Group"},
]
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Software Development :: Interpreters",
]
dynamic = ["version"]

[project.urls]
Homepage = "https://mathics.org/"

[project.optional-dependencies]
dev = [
    "pytest",
]
full = [
    "ujson",    # Optional Used in mathics_scanner.characters
]

[project.scripts]
mathics3-make-boxing-character-json = "mathics_scanner.generate.box_characters:main"
mathics3-make-named-character-json = "mathics_scanner.generate.named_characters:main"
mathics3-make-operator-json = "mathics_scanner.generate.operators:main"
mathics3-tokens = "mathics_scanner.mathics3_tokens:main"

[tool.setuptools]
packages = [
    "mathics_scanner",
    "mathics_scanner.data",
    "mathics_scanner.generate",
]

[tool.setuptools.package-data]
"mathics_scanner" = [
    "data/boxing-characters.json",
    "data/boxing-characters.yml",
    "data/named-characters.json",
    "data/named-characters.yml",
    "data/operators.yml",
    "data/operators.json",
    "data/*.csv",
    "data/*.json",
    "data/ExampleData/*",
]

[tool.setuptools.dynamic]
version = {attr = "mathics_scanner.version.__version__"}
